+++ src/redux/commonRecipes/commonSelectors.js
@@ -7,8 +7,3 @@ export const getRecipesBySearchQuery = state =>
   state.outerRecipes.recipesByQuery;
 
 export const getError = state => state.outerRecipes.isError;
-
-export const getRecipesListByIngredient = state =>
-  state.ingredients.recipesByIngredients;
-export const getIsIngredientsFetching = state =>
-  state.ingredients.isIngredientsFetching;
